Dual blob and buffer pause implementations (#2399)
* Update pause manager

* Implement base services for redis-only pause buffering

* Fix pause package tests

* Fix executor tests

* fix tests

* Add stub of PauseByID method

* comments

* Remove watermarks in favour of loading last block metadata

* Update tests

* fix race in mock tests

* Remove unused method in interface

* Add evaluable loader to manager interface

* Add Aggregated and BufferLen methods

Allows us to improve lookup perf with many events to choose whether to
immediately run expressions or aggregate.

* Simplify HandlePauses method in executor

* Allow empty event consumption, for eg. signals

* improve buffer consume

+++ pkg/execution/pauses/manager.go
@@ -3,9 +3,10 @@ package pauses
 import (
 	"context"
 	"errors"
-	"fmt"
 	"time"
 
+	"github.com/google/uuid"
+	"github.com/inngest/expr"
 	"github.com/inngest/inngest/pkg/execution/state"
 	"github.com/inngest/inngest/pkg/logger"
 )
@@ -14,9 +15,14 @@ var BlockFlushQueueName = "block-flush"
 
 var defaultFlushDelay = 10 * time.Second
 
-// StateBufferer transforms a state.Manager into a state.Bufferer
-func StateBufferer(rsm state.Manager) Bufferer {
-	return &redisAdapter{rsm}
+type ManagerOpt func(m Manager)
+
+func WithFlushDelay(delay time.Duration) ManagerOpt {
+	return func(m Manager) {
+		if mgr, ok := m.(*manager); ok {
+			mgr.flushDelay = delay
+		}
+	}
 }
 
 // NewManager returns a new pause writer, writing pauses to a Valkey/Redis/MemoryDB
@@ -24,13 +30,28 @@ func StateBufferer(rsm state.Manager) Bufferer {
 //
 // Blocks are flushed from the buffer in background jobs enqueued to the given queue.
 // This prevents eg. executors and new-runs from retaining blocks in-memory.
-func NewManager(buf Bufferer, bs BlockStore, flusher BlockFlushEnqueuer) *manager {
-	return &manager{
+func NewManager(buf Bufferer, bs BlockStore, flusher BlockFlushEnqueuer, opts ...ManagerOpt) Manager {
+	mgr := &manager{
 		buf:        buf,
 		bs:         bs,
 		flusher:    flusher,
 		flushDelay: defaultFlushDelay,
 	}
+
+	for _, o := range opts {
+		o(mgr)
+	}
+
+	return mgr
+}
+
+// NewRedisOnlyManager is a manager that only uses Redis as a buffer, without block flushing.
+func NewRedisOnlyManager(rsm state.PauseManager) Manager {
+	return NewManager(
+		StateBufferer(rsm),
+		nil,
+		nil,
+	)
 }
 
 type manager struct {
@@ -40,14 +61,50 @@ type manager struct {
 	flushDelay time.Duration
 }
 
-func (m manager) ConsumePause(ctx context.Context, pause state.Pause, opts state.ConsumePauseOpts) (state.ConsumePauseResult, func() error, error) {
-	if pause.Event == nil {
-		// A Pause must always have an event for this manager, else we cannot build the
-		// Index struct for deleting pauses.  It's also no longer possible to have pauses without
-		// events, so this should never happen.
-		return state.ConsumePauseResult{}, func() error { return nil }, fmt.Errorf("pause has no event")
+// PauseTimestamp returns the created at timestamp for a pause.
+func (m manager) PauseTimestamp(ctx context.Context, index Index, pause state.Pause) (time.Time, error) {
+	return m.buf.PauseTimestamp(ctx, index, pause)
+}
+
+func (m manager) PauseByInvokeCorrelationID(ctx context.Context, workspaceID uuid.UUID, correlationID string) (*state.Pause, error) {
+	return m.buf.PauseByInvokeCorrelationID(ctx, workspaceID, correlationID)
+}
+
+func (m manager) PauseBySignalID(ctx context.Context, workspaceID uuid.UUID, signal string) (*state.Pause, error) {
+	return m.buf.PauseBySignalID(ctx, workspaceID, signal)
+}
+
+func (m manager) BufferLen(ctx context.Context, idx Index) (int64, error) {
+	return m.buf.BufferLen(ctx, idx)
+}
+
+func (m manager) Aggregated(ctx context.Context, idx Index, minLen int64) (bool, error) {
+	// Check the buffer length by default.
+	n, err := m.buf.BufferLen(ctx, idx)
+	if err != nil {
+		return true, err
+	}
+	if n > minLen {
+		return true, nil
+	}
+	if m.bs == nil {
+		return false, nil
+	}
+	// If we've written a blob, aggregate, assuming there are always many pauses for this index.
+	return m.bs.IndexExists(ctx, idx)
+}
+
+func (m manager) IndexExists(ctx context.Context, i Index) (bool, error) {
+	ok, err := m.buf.IndexExists(ctx, i)
+	if err != nil || ok || m.bs == nil {
+		// It exists in the buffer, so no need to check blobstore.
+		return ok, err
 	}
 
+	return m.bs.IndexExists(ctx, i)
+}
+
+func (m manager) ConsumePause(ctx context.Context, pause state.Pause, opts state.ConsumePauseOpts) (state.ConsumePauseResult, func() error, error) {
 	// NOTE: There is a race condition when flushing blocks:  we may copy a pause
 	// into a block, then while writing the block to disk delete/consume a pause
 	// that is being written.  In this case the metadata for a block
@@ -67,11 +124,10 @@ func (m manager) ConsumePause(ctx context.Context, pause state.Pause, opts state
 	// case when consuming, and always re-delete the pause.  that’s no big deal, but
 	// not the best.
 	//
-	// In the future, we could add two block indexes:  pending, and stored.  this is a
+	// In the future, we could add two block indexes:  pending and flushed.  this is a
 	// pain, though, because we may die when uploading pending blocks, and that requires
 	// a bit of thought to work around, so we’ll just go with double deletes for now,
 	// assuming this won’t happen a ton.  this can be improved later.
-
 	res, cleanup, err := m.buf.ConsumePause(ctx, pause, opts)
 	// Is this an ErrDuplicateResponse?  If so, we've already consumed this pause,
 	// so delete it.  Similarly, if the error is nil we just consumed, so go ahead
@@ -80,9 +136,16 @@ func (m manager) ConsumePause(ctx context.Context, pause state.Pause, opts state
 		return res, cleanup, err
 	}
 
-	idx := Index{
-		pause.WorkspaceID,
-		*pause.Event,
+	idx := Index{WorkspaceID: pause.WorkspaceID}
+	if pause.Event != nil {
+		idx.EventName = *pause.Event
+	}
+
+	// Note that we cannot consume pauses from the blobstore with no event or backing
+	// blob.
+	if SkipFlushing(idx, []*state.Pause{&pause}) {
+		// This only exists in the buffer.  Return the buffer results.
+		return res, cleanup, err
 	}
 
 	// override the cleanup with idx deletion
@@ -114,15 +177,54 @@ func (m manager) Write(ctx context.Context, index Index, pauses ...*state.Pause)
 		return n, err
 	}
 
+	if m.bs == nil || SkipFlushing(index, pauses) {
+		// Don't bother flushing, as this needs to be kept in the buffer.
+		return n, nil
+	}
+
 	// If this is larger than the max buffer len, schedule a new block write.  We only
 	// enqueue this job once per index ID, using queue singletons to handle these.
-	if m.bs != nil && n >= m.bs.BlockSize() {
+	if n >= m.bs.BlockSize() {
 		if err := m.flusher.Enqueue(ctx, index); err != nil {
 			logger.StdlibLogger(ctx).Error("error attempting to flush block", "error", err)
 		}
 	}
 
 	return n, nil
+}
+
+func (m manager) PauseByID(ctx context.Context, index Index, pauseID uuid.UUID) (*state.Pause, error) {
+	// NOTE: This is only used to look up pauses when they time out.  As of this PR, timeout jobs
+	// embed each pause, prevent the need to do lookups.
+	//
+	// First, attempt to load this pause from the buffer.  Some pauses will definitely be here:
+	//
+	// - There aren't enough to flush to blocks, or we havent flushed yet.
+	// - We always keep pauses by ID for `step.invoke` and `step.waitForSignal` for fast O(1)
+	//   lookups to resolve these quickly
+	//
+	// If the pause isn't in the buffer, we check if the [env, event] index has been flushed before,
+	// and if so we attempt to load from the blobstore.
+	//
+	//
+	// # Loading from blobstores
+	//
+	// Loading pauses from the blobstore is hard. Pauses have V4 UUIDs as IDs:  they are random.
+	// This means there's no way of knowing which block/blob a pause belongs to without an index
+	// lookup of [pause ID] -> "created at".
+
+	pause, err := m.buf.PauseByID(ctx, index, pauseID)
+	if pause != nil && err == nil {
+		return pause, err
+	}
+
+	if m.bs != nil {
+		// We couldn't load from the buffer, so fall back.
+		return m.bs.PauseByID(ctx, index, pauseID)
+	}
+
+	// without a block store we should fall back to returning the error from the buffer.
+	return nil, err
 }
 
 // PausesSince loads pauses in the bfufer for a given index, since a given time.
@@ -134,6 +236,10 @@ func (m manager) PausesSince(ctx context.Context, index Index, since time.Time) 
 	bufIter, err := m.buf.PausesSince(ctx, index, since)
 	if err != nil {
 		return nil, err
+	}
+
+	if m.bs == nil {
+		return bufIter, nil
 	}
 
 	blocks, err := m.bs.BlocksSince(ctx, index, since)
@@ -150,9 +256,34 @@ func (m manager) PausesSince(ctx context.Context, index Index, since time.Time) 
 	), nil
 }
 
+// LoadEvaluablesSince calls PausesSince and implements the aggregate expression interface implementation
+// for grouping many pauses together.
+func (m manager) LoadEvaluablesSince(ctx context.Context, workspaceID uuid.UUID, eventName string, since time.Time, do func(context.Context, expr.Evaluable) error) error {
+	iter, err := m.PausesSince(ctx, Index{WorkspaceID: workspaceID, EventName: eventName}, since)
+	if err != nil {
+		return err
+	}
+
+	for iter.Next(ctx) {
+		pause := iter.Val(ctx)
+		if pause == nil {
+			continue
+		}
+		if err := do(ctx, pause); err != nil {
+			return err
+		}
+	}
+
+	if iter.Error() != context.Canceled && (iter.Error() != nil && iter.Error().Error() != "scan done") {
+		return iter.Error()
+	}
+
+	return nil
+}
+
 // Delete deletes a pause from from block storage or the buffer.
 func (m manager) Delete(ctx context.Context, index Index, pause state.Pause) error {
-	// XXX: Potential future optimization:  cache the last written block for an index
+	// Potential future optimization:  cache the last written block for an index
 	// in-memory so we can fast lookup here:
 	//
 	// if blockID.ts > pause.ts, skip deleting from the buffer as the pause is in a block.
@@ -163,12 +294,21 @@ func (m manager) Delete(ctx context.Context, index Index, pause state.Pause) err
 	if err != nil && !errors.Is(err, ErrNotInBuffer) {
 		return err
 	}
+
+	if m.bs == nil {
+		return nil
+	}
+
 	// Always also delegate to the flusher, just in case a block was written whilst
 	// we issued the delete request.
 	return m.bs.Delete(ctx, index, pause)
 }
 
 func (m manager) FlushIndexBlock(ctx context.Context, index Index) error {
+	if m.bs == nil {
+		return nil
+	}
+
 	// Ensure we delay writing the block.  This prevents clock skew on non-precision
 	// clocks from impacting out-of-order pauses;  we want pauses to be stored in-order
 	// and pause blocks to contain ordered pauses.
